+++ packages/media-parser/src/has-all-info.ts
@@ -1,4 +1,5 @@
 import {m3uHasStreams} from './containers/m3u/get-streams';
+import type {Options, ParseMediaFields} from './fields';
 import {hasAudioCodec} from './get-audio-codec';
 import {hasContainer} from './get-container';
 import {hasDimensions} from './get-dimensions';
@@ -11,7 +12,7 @@ import {hasSampleRate} from './get-sample-rate';
 import {getHasTracks} from './get-tracks';
 import {hasVideoCodec} from './get-video-codec';
 import {hasMetadata} from './metadata/get-metadata';
-import type {AllParseMediaFields, Options, ParseMediaFields} from './options';
+import type {AllParseMediaFields} from './options';
 import {maySkipVideoData} from './state/may-skip-video-data';
 import type {ParserState} from './state/parser-state';
