+++ packages/media-parser/src/parse-media-on-worker-entry.ts
@@ -1,10 +1,5 @@
-import type {
-	AllOptions,
-	Options,
-	ParseMediaFields,
-	ParseMediaOptions,
-	ParseMediaResult,
-} from './options';
+import type {AllOptions, Options, ParseMediaFields} from './fields';
+import type {ParseMediaOptions, ParseMediaResult} from './options';
 import type {OnAudioSample, OnVideoSample} from './webcodec-sample-types';
 import {deserializeError} from './worker/serialize-error';
 import type {
