+++ packages/media-parser/src/seeking-info.ts
@@ -1,0 +1,3 @@
+export type SeekingInfo = {
+	type: 'seeking-info';
+};
