+++ packages/media-parser/src/state/need-samples-for-fields.ts
@@ -1,4 +1,4 @@
-import type {AllOptions, Options, ParseMediaFields} from '../options';
+import type {AllOptions, Options, ParseMediaFields} from '../fields';
 
 const needsSamples: Record<keyof Options<ParseMediaFields>, boolean> = {
 	slowDurationInSeconds: true,
